Module Routes

Typed routing for OCaml. Routes provides combinators for adding typed routing to OCaml applications. The core library will be independent of any particular web framework or runtime.

module Method : sig ... end

type ('a, 'b) path

type 'b route

type 'b router

val int : ('a, 'b) path -> (int -> 'a, 'b) path

val int32 : ('a, 'b) path -> (int32 -> 'a, 'b) path

val int64 : ('a, 'b) path -> (int64 -> 'a, 'b) path

val str : ('a, 'b) path -> (string -> 'a, 'b) path

val bool : ('a, 'b) path -> (bool -> 'a, 'b) path

val s : string -> ('a, 'b) path -> ('a, 'b) path

val (/) : (('a, 'b) path -> 'c) -> ('d -> ('a, 'b) path) -> 'd -> 'c

val (/?) : (('a, 'b) path -> 'c) -> ('a, 'b) path -> 'c

val (@-->) : (unit -> ('a, 'b) path) -> 'a -> 'b route

val match' : ?⁠meth:Method.t -> 'a router -> target:string -> 'a option

val sprintf : (unit -> ('a, string) path) -> 'a

val nil : ('a, 'a) path

val one_of : (Method.t option * 'b route) list -> 'b router